Reorganized the project structure, added missing methods and access modifiers
