+++ Managers/ComponentManager.cpp
@@ -7,3 +7,11 @@
 void ComponentManager::addComponent(std::unique_ptr<IComponent> component) {
     components.push_front(std::unique_ptr<IComponent>(std::move(component)));
 }
+
+void ComponentManager::update(sf::Time& gameTime) {
+
+}
+
+void ComponentManager::draw(PtrRenderTarget target) {
+
+}
