+++ MeloRPG.cpp
@@ -4,6 +4,14 @@
 
 #include "MeloRPG.h"
 
+MeloRPG::MeloRPG() {
+
+}
+
+MeloRPG::~MeloRPG() {
+
+}
+
 void MeloRPG::createWindow(int width, int height, std::string title) {
     sf::VideoMode videoMode(width, height);
     _window = std::shared_ptr<sf::RenderWindow>(new sf::RenderWindow(videoMode, title));
@@ -29,10 +37,14 @@ void MeloRPG::update(sf::Time &game_time) {
             (event.type == sf::Event::KeyPressed && event.key.code == sf::Keyboard::Escape))
             _window->close();
     }
+
+    _componentManager.update(game_time);
 }
 
 void MeloRPG::draw() {
     _window->clear(sf::Color(100, 149, 237));
+
+    _componentManager.draw(_window);
 
     _window->display();
 }
